+++ test/eca/features/prompt_test.clj
@@ -2,7 +2,8 @@
   (:require
    [clojure.string :as string]
    [clojure.test :refer [deftest is testing]]
-   [eca.features.prompt :as prompt]))
+   [eca.features.prompt :as prompt]
+   [eca.test-helper :as h]))
 
 (deftest build-instructions-test
   (testing "Should create instructions with rules, contexts, and behavior"
@@ -14,7 +15,7 @@
                  {:name "rule2" :content "Second rule"}]
           fake-repo-map (delay "TREE")
           behavior "agent"
-          result (prompt/build-chat-instructions refined-contexts rules fake-repo-map behavior {})]
+          result (prompt/build-chat-instructions refined-contexts rules fake-repo-map behavior {} (h/db))]
       (is (string/includes? result "You are ECA"))
       (is (string/includes? result "<rules description=\"Rules defined by user\">"))
       (is (string/includes? result "<rule name=\"rule1\">First rule</rule>"))
@@ -35,7 +36,7 @@
                  {:name "rule2" :content "Second rule"}]
           fake-repo-map (delay "TREE")
           behavior "plan"
-          result (prompt/build-chat-instructions refined-contexts rules fake-repo-map behavior {})]
+          result (prompt/build-chat-instructions refined-contexts rules fake-repo-map behavior {} (h/db))]
       (is (string/includes? result "You are ECA"))
       (is (string/includes? result "<rules description=\"Rules defined by user\">"))
       (is (string/includes? result "<rule name=\"rule1\">First rule</rule>"))