+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix absolute paths being interpreted as commands. #199
 - Remove non used sync models code during initialize. #100
+- Fix system prompt to mention the user workspace roots.
 
 ## 0.78.4