+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.79.0
+
 - Fix absolute paths being interpreted as commands. #199
 - Remove non used sync models code during initialize. #100
 - Fix system prompt to mention the user workspace roots.